changed the applicationContext.xml to use transactions. (transaction annotations) added a specific library com.springsource.org.aopalliance-1.0.0.jar for that matter - because it wasnt present in the spring 3.0.1 we use.
OrderMapper maps also the customerID and companyID. customerID is just for testing purposes, when viewing all orders.

Annotated all the model beans.

The DAO implementations for Orders and Customers were annotated for transactions, as specified in the application context file.

The Orders REST service now gets JSON data for cars (DOES NOT RECOGNIZE array of cars) and gets QueryParameters in the link to specify order details from the customer.